--- project/src/mocks/reviews.ts
@@ -1,13 +1,0 @@
-export type ReviewType = {
-  comment: string;
-  date: string;
-  id?: number;
-  rating: number;
-  user: {
-    avatarUrl: string;
-    id: number;
-    isPro: boolean;
-    name: string;
-  };
-}
-
