+++ project/src/pages/offer-screen/offer-screen.tsx
@@ -2,13 +2,12 @@ import Logo from '../../components/logo/logo';
 import { Helmet } from 'react-helmet-async';
 import { CommentForm } from '../../components/comment-form/comment-form';
 import ReviewsList from '../../components/reviews-list/reviews-list';
-import { ReviewType } from '../../mocks/reviews';
+import { OfferType, ReviewType } from '../../types/state';
 import { useNavigate, useParams } from 'react-router-dom';
 import { useEffect, useState } from 'react';
 import { getRating} from '../../utils';
 import Map from '../../components/map/map';
 import Card from '../../components/card/card';
-import { OfferType } from '../../mocks/offers';
 import HeaderNav from '../../components/header-nav/header-nav';
 import { AppRoutes, AuthorizationsStatus, NameSpace } from '../../consts';
 import { useAppDispatch, useAppSelector } from '../../hooks';
